feat: add push to accounts connect modal
